requirements.txt
keras==2.9.0
Pillow==9.0.1
streamlit==1.10.0
streamlit-option-menu==0.3.2
tensorflow==2.9.1
pandas<1.2.0
plotly==4.14.3
opencv-python==3.4.2.16

